Amélioration du parcours copro et différenciation des composants AddressSearch et LogementSearch
