+++ app/copropriete/simulationConfigCopro.yaml
@@ -7,5 +7,3 @@ prioritaires:
   - copropriété . gain énergétique
   - copropriété . bonus sortie passoire
   - copropriété . bonus copropriété fragile
-non prioritaires:
-  -
