+++ components/publicodes/getNextQuestions.ts
@@ -6,10 +6,14 @@ export default function getNextQuestions(
   questionsConfig = [],
 ) {
   const { missingVariables } = evaluation
-  if (questionsConfig?.prioritaires?.length > 0)
-    return questionsConfig.prioritaires
-      .filter((question) => Object.keys(missingVariables).includes(question))
-      .filter(([question]) => !answeredQuestions.includes(question))
+  if (questionsConfig?.prioritaires?.length > 0) {
+    let questions = questionsConfig.prioritaires.filter((question) =>
+      Object.keys(missingVariables).includes(question),
+    )
+    questions.unshift(...(questionsConfig?.préface || []))
+    return questions.filter((question) => !answeredQuestions.includes(question))
+  }
+
   const allMissingEntries = Object.entries(missingVariables),
     missingEntries = allMissingEntries.filter(
       ([question]) => !answeredQuestions.includes(question),
